Merge Development branch into Main Branch after Sprint 4

+++ backend/Api/Module/Jobs/Jobs.Application/JobApplication/Delete/v1/DeleteJobApplicationHandler.cs
@@ -4,6 +4,7 @@ using TalentMesh.Module.Job.Domain.Exceptions;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using TalentMesh.Module.Job.Domain.Extentsion;
 
 namespace TalentMesh.Module.Job.Application.JobApplication.Delete.v1;
 public sealed class DeleteJobApplicationHandler(
@@ -15,7 +16,7 @@ public sealed class DeleteJobApplicationHandler(
     {
         ArgumentNullException.ThrowIfNull(request);
         var jobApplication = await repository.GetByIdAsync(request.Id, cancellationToken);
-        _ = jobApplication ?? throw new JobApplicationNotFoundException(request.Id);
+        if (jobApplication.IsDeletedOrNotFound()) throw new JobNotFoundException(request.Id);
         await repository.DeleteAsync(jobApplication, cancellationToken);
         logger.LogInformation("JobApplication with id : {Id} deleted", jobApplication.Id);
     }
